requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
ultralytics
opencv-python-headless
python-dotenv